Fixed Test Cases for TaskController and TaskService, and refactored code for TaskController and TaskService
