+++ src/main/java/edu/tcu/cs/peerevalutationtool/repository/TaskRepository.java
@@ -1,6 +1,5 @@
 package edu.tcu.cs.peerevalutationtool.repository;
 
-import edu.tcu.cs.peerevalutationtool.domain.Team;
 import edu.tcu.cs.peerevalutationtool.task.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -11,5 +10,5 @@ import java.util.List;
 public interface TaskRepository extends JpaRepository<Task, Long> {
     List<Task> findAllByStudentId(Long studentId);
     List<Task> findAllByWeek(String taskWeek);
-    List<Task> findAllByWeekAndStudentTeamId(String taskWeek, long taskTeamId);
+    List<Task> findAllByWeekAndStudentTeamId(String taskWeek, String taskTeamId);
 }
